+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Enable/disable MEMS self-test function.
 - Support interrupts.
+- Support physical interrupt pin configuration.
 
 ### Changed
 - [breaking-change] Renamed communication self-test function `self_test()` to